Fix warnings.

+++ include/gtirb/Utilities.hpp
@@ -95,7 +95,7 @@ namespace gtirb
 
             source.u = u;
 
-            for(int k = 0; k < sizeof(T); k++)
+            for(size_t k = 0; k < sizeof(T); k++)
             {
                 dest.u8[k] = source.u8[sizeof(T) - k - 1];
             }
